Renamed to page-item-authoring
