+++ app/assets/javascripts/application.js
@@ -28,7 +28,7 @@
 //= require check-answer
 //= require lara-plugin-api-V2
 //= require lara-typescript
-//= require page-item-admin
+//= require page-item-authoring
 //= require ./common.js
 //= require edit_sequence
 //= require login_menu
