+++ lara-typescript/package.json
@@ -5,12 +5,12 @@
   "scripts": {
     "publish:plugin-api": "npm run build && cd dist/plugin-api && npm publish --access public",
     "build": "npm-run-all lint:build clean build:webpack build:docs",
-    "build:webpack": "webpack --devtool false && cp ./dist/lara-typescript.js ../app/assets/javascripts/ && cp ./dist/lara-typescript.css ../app/assets/stylesheets/ && cp ./dist/page-item-admin.js ../app/assets/javascripts/",
+    "build:webpack": "webpack --devtool false && cp ./dist/lara-typescript.js ../app/assets/javascripts/ && cp ./dist/lara-typescript.css ../app/assets/stylesheets/ && cp ./dist/page-item-authoring.js ../app/assets/javascripts/",
     "build:docs": "npm-run-all build:docs:plugin-api build:docs:fix-doc-paths build:docs:copy-files",
     "build:docs:plugin-api": "typedoc --theme markdown --readme src/plugin-api/README.md --ignoreCompilerErrors --name 'LARA Plugin API' --excludeNotExported --excludeExternals --excludePrivate --mode file --out ./docs/lara-plugin-api src/plugin-api",
     "build:docs:fix-doc-paths": "./fix-doc-paths.js",
     "build:docs:copy-files": "cpx './docs/lara-plugin-api/**/*.*' ../docs/lara-plugin-api/",
-    "build:dev:page-item-admin": "webpack && cp ./dist/page-item-admin.js ../app/assets/javascripts/",
+    "build:dev:page-item-authoring": "webpack && cp ./dist/page-item-authoring.js ../app/assets/javascripts/",
     "clean": "rimraf dist docs ../docs/lara-plugin-api",
     "lint": "tslint -c tslint.json 'src/**/*.{ts,tsx}'",
     "lint:build": "tslint -c tslint-build.json 'src/**/*.{ts,tsx}'",
